+++ demos/26-girih/GirihBowtie.js
@@ -166,11 +166,13 @@ GirihBowtie.prototype.getOuterTilePolygonAt = GirihTile.prototype.getOuterTilePo
 GirihBowtie.prototype.getTranslatedVertex   = GirihTile.prototype.getTranslatedVertex;
 //GirihBowtie.prototype.containsPoint         = GirihTile.prototype.containsPoint;
 GirihBowtie.prototype.containsVert         = Polygon.prototype.containsVert;
+GirihBowtie.prototype.getBounds         = Polygon.prototype.getBounds;
 GirihBowtie.prototype.rotate         = GirihTile.prototype.rotate;
 GirihBowtie.prototype.locateEdgeAtPoint     = GirihTile.prototype.locateEdgeAtPoint;
 GirihBowtie.prototype.locateAdjacentEdge    = GirihTile.prototype.locateAdjacentEdge;
 GirihBowtie.prototype.getVertexAt           = Polygon.prototype.getVertexAt; // GirihTile.prototype.getVertexAt;
-GirihBowtie.prototype.toSVG                 = GirihTile.prototype.toSVG;
+// GirihBowtie.prototype.toSVG                 = GirihTile.prototype.toSVG;
+GirihBowtie.prototype.toSVGString                 = Polygon.prototype.toSVGString;
 GirihBowtie.prototype.move         = GirihTile.prototype.move;
 GirihBowtie.prototype.findAdjacentTilePosition = GirihTile.prototype.findAdjacentTilePosition;